chat fix

+++ back/app/Http/Controllers/ChatController.php
@@ -50,11 +50,14 @@ class ChatController extends Controller
     }
 
     public function getChats () {
-        $chats = DB::table('chats')
-        ->join('users', 'users.id', '=', 'chats.idusuario')
-        ->select('users.name as user', 'chats.id', 'chats.message', 'chats.type')
-        ->orderByDesc('id')->get();
-        return response()->json(['chats' => $chats]);
+        $user = Auth::user();
+        if ($user->rank > 2) {
+            $chats = DB::table('chats')
+            ->join('users', 'users.id', '=', 'chats.idusuario')
+            ->select('users.name as user', 'chats.id', 'chats.message', 'chats.type')
+            ->orderByDesc('id')->get();
+            return response()->json(['chats' => $chats]);
+        }
     }
 
     public function getChatById ($id) {
@@ -66,27 +69,33 @@ class ChatController extends Controller
     }
 
     public function deleteChatById ($id) {
-        $chat = DB::table('chats')->where('id', '=', $id)->delete();
-        return response()->json(['chat' => $chat]);
+        $user = Auth::user();
+        if ($user->rank > 2) {
+            $chat = DB::table('chats')->where('id', '=', $id)->delete();
+            return response()->json(['chat' => $chat]);
+        }
     }
 
     public function answerChat (Request $request) {
-        $id = $request->input('id');
-        $answer = $request->input('answer');
         $user = Auth::user();
-        $chat = DB::table('chats')
-        ->join('users', 'users.id', '=', 'chats.idusuario')
-        ->select('users.email as email', 'chats.id', 'chats.message', 'chats.type')
-        ->where('chats.id', '=', $id)
-        ->orderByDesc('id')->get()->first();
-        $title = 'Respuesta a su '.$chat->type;
-        $body = $answer;
-        $body .= "<br>";
-        $body .= "<br>";
-        $body .= "Att:";
-        $body .= $user->name;
-        Mail::to($chat->email)->send(new MessageSend($title,$body,$chat->email));
-        return response()->json(['chat' => $chat, 'message' => 'CREATED'], 201);
+        if ($user->rank > 2) {
+            $id = $request->input('id');
+            $answer = $request->input('answer');
+            $user = Auth::user();
+            $chat = DB::table('chats')
+            ->join('users', 'users.id', '=', 'chats.idusuario')
+            ->select('users.email as email', 'chats.id', 'chats.message', 'chats.type')
+            ->where('chats.id', '=', $id)
+            ->orderByDesc('id')->get()->first();
+            $title = 'Respuesta a su '.$chat->type;
+            $body = $answer;
+            $body .= "<br>";
+            $body .= "<br>";
+            $body .= "Att:";
+            $body .= $user->name;
+            Mail::to($chat->email)->send(new MessageSend($title,$body,$chat->email));
+            return response()->json(['chat' => $chat, 'message' => 'CREATED'], 201);
+        }
     }
